moved function for converting MS to Seconds to utils
