+++ eml/services/utilityFunctions.js
@@ -13,6 +13,18 @@ export function getDifficultyLabel(lvl) {
   }
 };
 
+
+export const convertMsToTime = (ms) => {
+  let seconds = Math.floor((ms / 1000) % 60);
+  let minutes = Math.floor((ms / (1000 * 60)) % 60);
+
+  seconds = seconds < 10 ? '0' + seconds : seconds;
+  minutes = minutes < 10 ? '0' + minutes : minutes;
+
+  return `${minutes}:${seconds}`;
+}
+
+
 export function determineCategory(category) {
   switch (category) {
     case "personal finance":
